perf(rn): defer react-native-web off the jobboard eager bundle
Two eager importers were dragging all of react-native-web into the
landing chunk:

1. KbveProvider imported AppState (react-native) for foreground
   auto-refresh and wrapped children in SafeAreaProvider. Split both into
   platform helpers — useAutoRefresh (.web uses document visibilitychange)
   and AuthSafeArea (.web is a passthrough) — so the shared KbveProvider
   pulls no react-native.
2. main.tsx imported 'react-native-url-polyfill/auto', whose only purpose
   is `import {Platform} from 'react-native'` + a Platform.OS!=='web'
   check (a no-op on web). Browsers have native URL, so drop it.

react-native-web now loads only with the routes that use it (dashboard,
login). Eager bundle no longer contains any react-native/reanimated/expo
module (sourcemap-verified); landing/browse visitors never download it.
