+++ packages/npm/rn/src/auth/AuthSafeArea.tsx
@@ -1,0 +1,6 @@
+import type { ReactNode } from 'react';
+import { SafeAreaProvider } from 'react-native-safe-area-context';
+
+export function AuthSafeArea({ children }: { children: ReactNode }) {
+	return <SafeAreaProvider>{children}</SafeAreaProvider>;
+}
