+++ packages/npm/rn/src/auth/AuthSafeArea.web.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react';
+
+export function AuthSafeArea({ children }: { children: ReactNode }) {
+	return <>{children}</>;
+}
